Split native Zenmap scan comparison helpers
